Continue event monitoring after end until new instance
