+++ packages/playground/src/data/saga.js
@@ -1,5 +1,7 @@
+import setSpeechRecognitionInstance from './sagas/setSpeechRecognitionInstance';
 import startSpeechRecognition from './sagas/startSpeechRecognition';
 
 export default function* () {
+  yield* setSpeechRecognitionInstance();
   yield* startSpeechRecognition();
 }
