+++ packages/playground/src/data/sagas/effects/forever.js
@@ -1,5 +1,5 @@
 import { call } from 'redux-saga/effects';
 
 export default function* () {
-  yield call(new Promise(() => {}));
+  yield call(() => new Promise(() => {}));
 }
